added typecasts

+++ src/joystick_control.cpp
@@ -276,11 +276,11 @@ void joy_callback(const sensor_msgs::Joy::ConstPtr &msg){
 
     printf("%f, %f, %f, %f, %f, %f, %f, %f\n", msg->axes[0],msg->axes[1],msg->axes[2],msg->axes[3],msg->axes[4],msg->axes[5],msg->axes[6],msg->axes[7]);
     printf("FB: %f LR: %f YAW: %f PITCH: %f\n\n",msg->axes[gamepad.FB], msg->axes[gamepad.LR], msg->axes[gamepad.YAW], msg->axes[gamepad.PITCH]);
-    cmd.velocity[0] = msg->axes[gamepad.FB]/gamepad.resolution * linear_speed_unit;
-    cmd.velocity[1] = msg->axes[gamepad.LR]/gamepad.resolution * linear_speed_unit;
+    cmd.velocity[0] = ((float) msg->axes[gamepad.FB])/gamepad.resolution * linear_speed_unit;
+    cmd.velocity[1] = ((float) msg->axes[gamepad.LR])/gamepad.resolution * linear_speed_unit;
 
-    cmd.yawSpeed =  msg->axes[gamepad.YAW]/gamepad.resolution * rotational_speed_unit;
-    cmd.euler[1] = msg->axes[gamepad.PITCH]/gamepad.resolution * angle_unit;
+    cmd.yawSpeed =  ((float) msg->axes[gamepad.YAW])/gamepad.resolution * rotational_speed_unit;
+    cmd.euler[1] = ((float) msg->axes[gamepad.PITCH])/gamepad.resolution * angle_unit;
 
     update_buttons(msg);
